+++ narwhals/_plan/dummy.py
@@ -15,7 +15,8 @@ from narwhals._plan import (
     functions as F,
     operators as ops,
 )
-from narwhals._plan.common import NamedIR, into_dtype, is_column, is_expr, is_series
+from narwhals._plan._guards import is_column, is_expr, is_series
+from narwhals._plan.common import NamedIR, into_dtype
 from narwhals._plan.contexts import ExprContext
 from narwhals._plan.options import (
     EWMOptions,
